pyproject.toml
[tool.poetry]
name = "repl_nix_banana-milk"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
discord = "^2.1.0"
pylast = "^5.1.0"
requests-cache = "^0.9.7"
ipython = "^8.6.0"
pandas = "^1.5.2"
mysql-connector-python = "^8.0.31"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
